requirements.txt
feedparser
requests
beautifulsoup4
google-auth
gspread
gTTS
PyYAML
google-generativeai
